requirements.txt
annotated-types
attrs
blinker
cachelib
certifi
charset-normalizer
ci-info
click
colorama
configobj
configparser
defusedxml
docstring_parser
docx
etelemetry
filelock
fire
Flask
Flask-Session
fsspec
gevent
httplib2
huggingface-hub==0.24.7
idna
isodate
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
looseversion
lxml
MarkupSafe
mpmath
networkx
nibabel
nipype
numpy
odfpy
packaging
pandas
pathlib
pdf2image
pillow
prov
pybind11
pydantic
pydantic_core
pydot
PyMuPDF
PyMuPDFb
pyparsing
PyPDF2
pytesseract
python-dateutil
python-docx
python-dotenv
pytz
pyxnat
PyYAML
rdflib
referencing
regex
requests
rpds-py
safetensors
scipy
sentencepiece
setuptools
simple_parsing
simplejson
six
sympy
spacy
termcolor
tika
tiktoken
tokenizers
tqdm
traits
transformers
typing_extensions
tzdata
urllib3
Werkzeug
waitress
huggingface-hub
striprtf
easyocr
gunicorn
